refactor: move utility functions to util.go
